Move xdrposix.c to xdr.c
Also hard-codes xdrposix calls for op-specific calls
