+++ mfhdf/libsrc/CMakeLists.txt
@@ -32,7 +32,6 @@ set (HDF4_MFHDF_LIBSRC_CSRCS
     ${HDF4_MFHDF_LIBSRC_SOURCE_DIR}/string.c
     ${HDF4_MFHDF_LIBSRC_SOURCE_DIR}/var.c
     ${HDF4_MFHDF_LIBSRC_SOURCE_DIR}/xdr.c
-    ${HDF4_MFHDF_LIBSRC_SOURCE_DIR}/xdrposix.c
 )
 
 set (HDF4_MFHDF_LIBSRC_INSTALL_CHDRS
